draft of the website
mostly copied data from my old website, converting files from org to
markdown.

was able to re-use the css wholesale, which is great.  and not
unexpected.

I decided to store talks and papers as YAML databases in the _data/
directory.  hopefully this proves easier to maintain...

I haven't touched the content at all, so that needs to be updated